add new allocation status for scratch orgs in pool

+++ src/impl/pool/scratchorg/PoolDeleteImpl.ts
@@ -7,22 +7,26 @@ export default class PoolDeleteImpl {
   private tag: string;
   private mypool: boolean;
   private allScratchOrgs: boolean;
+  private inprogressonly: boolean;
 
   public constructor(
     hubOrg: Org,
     apiversion: string,
     tag: string,
     mypool: boolean,
-    allScratchOrgs: boolean
+    allScratchOrgs: boolean,
+    inprogressonly: boolean
   ) {
     this.hubOrg = hubOrg;
     this.apiversion = apiversion;
     this.tag = tag;
     this.mypool = mypool;
     this.allScratchOrgs = allScratchOrgs;
+    this.inprogressonly = inprogressonly;
   }
 
   public async execute(): Promise<ScratchOrg[]> {
+    await ScratchOrgUtils.checkForNewVersionCompatible(this.hubOrg);
     const results = (await ScratchOrgUtils.getScratchOrgsByTag(
       this.tag,
       this.hubOrg,
@@ -33,40 +37,47 @@ export default class PoolDeleteImpl {
     let scratchOrgToDelete: ScratchOrg[] = new Array<ScratchOrg>();
     if (results.records.length > 0) {
       SFPowerkit.log(
-        `${this.tag} pool has ${results.records.length} Scratch orgs available`,
+        `${this.tag} pool has ${results.records.length} Scratch orgs.`,
         LoggerLevel.TRACE
       );
 
       let scrathOrgIds: string[] = [];
 
       for (let element of results.records) {
-        let soDetail: ScratchOrg = {};
-        soDetail.orgId = element.ScratchOrg;
-        soDetail.loginURL = element.LoginUrl;
-        soDetail.username = element.SignupUsername;
-        soDetail.expityDate = element.ExpirationDate;
-        soDetail.status = "Deleted";
+        if (
+          !this.inprogressonly ||
+          element.Allocation_status__c === "In Progress"
+        ) {
+          let soDetail: ScratchOrg = {};
+          soDetail.orgId = element.ScratchOrg;
+          soDetail.loginURL = element.LoginUrl;
+          soDetail.username = element.SignupUsername;
+          soDetail.expityDate = element.ExpirationDate;
+          soDetail.status = "Deleted";
 
-        scratchOrgToDelete.push(soDetail);
-        scrathOrgIds.push(`'${element.Id}'`);
+          scratchOrgToDelete.push(soDetail);
+          scrathOrgIds.push(`'${element.Id}'`);
+        }
       }
 
-      let activeScrathOrgs = await ScratchOrgUtils.getActiveScratchOrgsByInfoId(
-        this.hubOrg,
-        scrathOrgIds.join(",")
-      );
+      if (scrathOrgIds.length > 0) {
+        let activeScrathOrgs = await ScratchOrgUtils.getActiveScratchOrgsByInfoId(
+          this.hubOrg,
+          scrathOrgIds.join(",")
+        );
 
-      if (activeScrathOrgs.records.length > 0) {
-        for (let ScratchOrg of activeScrathOrgs.records) {
-          await ScratchOrgUtils.deleteScratchOrg(
-            this.hubOrg,
-            this.apiversion,
-            ScratchOrg.Id
-          );
-          SFPowerkit.log(
-            `Scratch org with username ${ScratchOrg.SignupUsername} is deleted successfully`,
-            LoggerLevel.TRACE
-          );
+        if (activeScrathOrgs.records.length > 0) {
+          for (let ScratchOrg of activeScrathOrgs.records) {
+            await ScratchOrgUtils.deleteScratchOrg(
+              this.hubOrg,
+              this.apiversion,
+              ScratchOrg.Id
+            );
+            SFPowerkit.log(
+              `Scratch org with username ${ScratchOrg.SignupUsername} is deleted successfully`,
+              LoggerLevel.TRACE
+            );
+          }
         }
       }
     }
